Unmodified version of cherrypy.wsgiserver as downloaded from http://download.cherrypy.org/cherrypy/3.1.2/
